Improve relationship normalization
Normalize currently uses the alphabetical order to normalize a two-way
relationship. Until now, one-way relationships were not properly
handled. The alphabetical order would sometimes make Normalize return
a relationship whose FromName field was set to an empty string. This
itself was not a huge problem, but it is cleaner and makes more sense
to use the original relationship when it does not point to another
type's relationship (one-way) than to use its inverse.